refactor: group analytics files into sub folders
